requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-JWT-Extended==4.1.0
flask-marshmallow==0.14.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
flasgger==0.9.5
webargs==7.0.1
flake8
black